+++ cligen_object.c
@@ -388,103 +388,102 @@ cov_pref(cg_obj *co)
 
     switch (co->co_vtype){
     case CGV_ERR:
-        pref = 0; /* Illegal */
+        pref = COV_PREF_ERROR; /* Illegal */
         break;
         /* ints in range 22-60 */
     case CGV_INT8:
         if (co->co_rangelen)
-            pref = 60;
+            pref = COV_PREF_INT8_RANGE;
         else
-            pref = 52;
+            pref = COV_PREF_INT8;
         break;
     case CGV_INT16:
         if (co->co_rangelen)
-            pref = 58;
+            pref = COV_PREF_INT16_RANGE;
         else
-            pref = 50;
+            pref = COV_PREF_INT16;
         break;
     case CGV_INT32:
         if (co->co_rangelen)
-            pref = 56;
+            pref = COV_PREF_INT32_RANGE;
         else
-            pref = 48;
+            pref = COV_PREF_INT32;
         break;
     case CGV_INT64:
         if (co->co_rangelen)
-            pref = 54;
+            pref = COV_PREF_INT64_RANGE;
         else
-            pref = 46;
+            pref = COV_PREF_INT64;
         break;
     case CGV_UINT8:
         if (co->co_rangelen)
-            pref = 59;
+            pref = COV_PREF_UINT8_RANGE;
         else
-            pref = 51;
+            pref = COV_PREF_UINT8;
         break;
     case CGV_UINT16:
         if (co->co_rangelen)
-            pref = 57;
+            pref = COV_PREF_UINT16_RANGE;
         else
-            pref = 49;
+            pref = COV_PREF_UINT16;
         break;
     case CGV_UINT32:
         if (co->co_rangelen)
-            pref = 55;
+            pref = COV_PREF_UINT32_RANGE;
         else
-            pref = 47;
+            pref = COV_PREF_UINT32;
         break;
     case CGV_UINT64:
         if (co->co_rangelen)
-            pref = 53;
+            pref = COV_PREF_UINT64_RANGE;
         else
-            pref = 45;
+            pref = COV_PREF_UINT64;
         break;
     case CGV_DEC64:
-        pref = 62;
+        pref = COV_PREF_DECIMAL64;
         break;
     case CGV_BOOL:
-        pref = 12;
+        pref = COV_PREF_BOOL;
         break;
     case CGV_REST:
-        pref = 1;
+        pref = COV_PREF_REST;
         break;
     case CGV_STRING:
         if (co->co_expand_fn_str != NULL)
-            pref = 8;
+            pref = COV_PREF_STRING_EXPAND;
         else if (co->co_regex)
-            pref = 7;
+            pref = COV_PREF_STRING_REGEXP;
         else
-            pref = 5;
+            pref = COV_PREF_STRING;
         break;
     case CGV_INTERFACE:
-        pref = 10;
+        pref = COV_PREF_INTERFACE;
         break;
     case CGV_IPV4ADDR:
     case CGV_IPV4PFX:
-        pref = 70;
+        pref = COV_PREF_IPV4ADDR;
         break;
     case CGV_IPV6ADDR:
     case CGV_IPV6PFX:
-        pref = 71;
+        pref = COV_PREF_IPV6ADDR;
         break;
     case CGV_MACADDR:
-        pref = 72;
+        pref = COV_PREF_MACADDR;
         break;
     case CGV_URL:
-        pref = 20;
+        pref = COV_PREF_URL;
         break;
     case CGV_UUID:
-        pref = 73;
+        pref = COV_PREF_UUID;
         break;
     case CGV_TIME:
-        pref = 74;
+        pref = COV_PREF_TIME;
         break;
     case CGV_VOID: /* N/A */
         break;
     case CGV_EMPTY:
         break;
     }
-
     return pref;
 }